BlockGraphIterator, BlockGraphIteratorAndTrapAggregator separate classes, buildtraps to LabeledStmtPrinter
