+++ sootup.core/src/main/java/sootup/core/util/printer/JimplePrinter.java
@@ -301,44 +301,6 @@ public class JimplePrinter {
   }
 
   /**
-   * returns a (reconstructed) list of traps like the traptable in the bytecode
-   *
-   * <p>Note: if you need exceptionional flow information in more augmented with the affected
-   * blocks/stmts and not just a (reconstructed, possibly more verbose) traptable - have a look at
-   * BasicBlock.getExceptionalSuccessor()
-   */
-  /** hint: little expensive getter - its more of a build/create - currently no overlaps */
-  public List<Trap> buildTraps(StmtGraph stmtGraph) {
-    // [ms] try to incorporate it into the serialisation of jimple printing so the other half of
-    // iteration information is not wasted..
-    StmtGraph<MutableBasicBlock>.BlockGraphIteratorAndTrapAggregator it = stmtGraph.new BlockGraphIteratorAndTrapAggregator(new MutableBasicBlockImpl());
-    // it.getTraps() is valid/completely build when the iterator is done.
-    Map<Stmt, Integer> stmtsBlockIdx = new IdentityHashMap<>();
-    int i = 0;
-    // collect BlockIdx positions to sort the traps according to the numbering
-    while (it.hasNext()) {
-      final BasicBlock<?> nextBlock = it.next();
-      stmtsBlockIdx.put(nextBlock.getHead(), i);
-      stmtsBlockIdx.put(nextBlock.getTail(), i);
-      i++;
-    }
-    final List<Trap> traps = it.getTraps();
-    traps.sort(getTrapComparator(stmtsBlockIdx));
-    return traps;
-  }
-
-  /** Comparator which sorts the trap output in getTraps() */
-  public Comparator<Trap> getTrapComparator(@Nonnull Map<Stmt, Integer> stmtsBlockIdx) {
-    return (a, b) ->
-            ComparisonChain.start()
-                    .compare(stmtsBlockIdx.get(a.getBeginStmt()), stmtsBlockIdx.get(b.getBeginStmt()))
-                    .compare(stmtsBlockIdx.get(a.getEndStmt()), stmtsBlockIdx.get(b.getEndStmt()))
-                    // [ms] would be nice to have the traps ordered by exception hierarchy as well
-                    .compare(a.getExceptionType().toString(), b.getExceptionType().toString())
-                    .result();
-  }
-
-  /**
    * Prints out the method corresponding to b Body, (declaration and body), in the textual format
    * corresponding to the IR used to encode b body.
    *
@@ -421,7 +383,7 @@ public class JimplePrinter {
 
     // Print out exceptions
     {
-      Iterator<Trap> trapIt = buildTraps(stmtGraph).iterator();
+      Iterator<Trap> trapIt = printer.buildTraps(stmtGraph).iterator();
 
       if (trapIt.hasNext()) {
         printer.newline();
