+++ sootup.core/src/main/java/sootup/core/util/printer/LabeledStmtPrinter.java
@@ -24,11 +24,15 @@ package sootup.core.util.printer;
 
 import java.util.*;
 import javax.annotation.Nonnull;
+
+import com.google.common.collect.ComparisonChain;
+import sootup.core.graph.BasicBlock;
 import sootup.core.graph.StmtGraph;
 import sootup.core.jimple.Jimple;
 import sootup.core.jimple.basic.Trap;
 import sootup.core.jimple.common.ref.IdentityRef;
-import sootup.core.jimple.common.stmt.Stmt;
+import sootup.core.jimple.common.stmt.*;
+import sootup.core.jimple.javabytecode.stmt.JSwitchStmt;
 import sootup.core.model.SootField;
 import sootup.core.model.SootMethod;
 import sootup.core.signatures.FieldSignature;
@@ -113,8 +117,8 @@ public abstract class LabeledStmtPrinter extends AbstractStmtPrinter {
 
   @Nonnull
   public List<Stmt> getStmts(@Nonnull StmtGraph<?> stmtGraph) {
-    final Collection<Stmt> targetStmtsOfBranches = stmtGraph.getLabeledStmts();
-    final List<Trap> traps = new JimplePrinter().buildTraps(stmtGraph);
+    final Collection<Stmt> targetStmtsOfBranches = getLabeledStmts(stmtGraph);
+    final List<Trap> traps = buildTraps(stmtGraph);
 
     final int maxEstimatedSize = targetStmtsOfBranches.size() + traps.size() * 3;
     labels = new HashMap<>(maxEstimatedSize, 1);
@@ -136,7 +140,7 @@ public abstract class LabeledStmtPrinter extends AbstractStmtPrinter {
     // or is the begin of a trap-range) or does it mark the end of a trap range
     // does it need a label
     for (Stmt stmt : targetStmtsOfBranches) {
-      if (stmtGraph.isStmtBranchTarget(stmt) || trapStmts.contains(stmt)) {
+      if ((stmt instanceof JNopStmt) || stmtGraph.isStmtBranchTarget(stmt) || trapStmts.contains(stmt)) {
         labelStmts.add(stmt);
       } else {
         refStmts.add(stmt);
@@ -194,5 +198,82 @@ public abstract class LabeledStmtPrinter extends AbstractStmtPrinter {
     final FieldSubSignature subSignature = fieldSig.getSubSignature();
     typeSignature(subSignature.getType());
     output.append(' ').append(Jimple.escape(subSignature.getName())).append('>');
+  }
+
+  /**
+   * returns a (reconstructed) list of traps like the traptable in the bytecode
+   *
+   * <p>Note: if you need exceptionional flow information in more augmented with the affected
+   * blocks/stmts and not just a (reconstructed, possibly more verbose) traptable - have a look at
+   * BasicBlock.getExceptionalSuccessor()
+   */
+  /** hint: little expensive getter - its more of a build/create - currently no overlaps */
+  public List<Trap> buildTraps(StmtGraph stmtGraph) {
+    // [ms] try to incorporate it into the serialisation of jimple printing so the other half of
+    // iteration information is not wasted..
+    BlockGraphIteratorAndTrapAggregator it = new BlockGraphIteratorAndTrapAggregator(stmtGraph);
+    // it.getTraps() is valid/completely build when the iterator is done.
+    Map<Stmt, Integer> stmtsBlockIdx = new IdentityHashMap<>();
+    int i = 0;
+    // collect BlockIdx positions to sort the traps according to the numbering
+    while (it.hasNext()) {
+      final BasicBlock<?> nextBlock = it.next();
+      stmtsBlockIdx.put(nextBlock.getHead(), i);
+      stmtsBlockIdx.put(nextBlock.getTail(), i);
+      i++;
+    }
+    final List<Trap> traps = it.getTraps();
+    if (it.getLastStmt() != null) {
+      stmtsBlockIdx.put(it.getLastStmt(), i);
+    }
+    traps.sort(getTrapComparator(stmtsBlockIdx));
+    return traps;
+  }
+
+  /** Comparator which sorts the trap output in getTraps() */
+  public Comparator<Trap> getTrapComparator(@Nonnull Map<Stmt, Integer> stmtsBlockIdx) {
+    return (a, b) ->
+            ComparisonChain.start()
+                    .compare(stmtsBlockIdx.get(a.getBeginStmt()), stmtsBlockIdx.get(b.getBeginStmt()))
+                    .compare(stmtsBlockIdx.get(a.getEndStmt()), stmtsBlockIdx.get(b.getEndStmt()))
+                    // [ms] would be nice to have the traps ordered by exception hierarchy as well
+                    .compare(a.getExceptionType().toString(), b.getExceptionType().toString())
+                    .result();
+  }
+
+  /**
+   * Returns the result of iterating through all Stmts in this body. All Stmts thus found are
+   * returned. Branching Stmts and statements which use PhiExpr will have Stmts; a Stmt contains a
+   * Stmt that is either a target of a branch or is being used as a pointer to the end of a CFG
+   * block.
+   *
+   * <p>This method was typically used for pointer patching, e.g. when the unit chain is cloned.
+   *
+   * @return A collection of all the Stmts that are targets of a BranchingStmt
+   */
+  @Nonnull
+  public Collection<Stmt> getLabeledStmts(StmtGraph stmtGraph) {
+    Set<Stmt> stmtList = new HashSet<>();
+    Collection<Stmt> stmtGraphNodes = stmtGraph.getNodes();
+    for (Stmt stmt : stmtGraphNodes) {
+      if (stmt instanceof BranchingStmt) {
+        if (stmt instanceof JIfStmt) {
+          stmtList.add((Stmt) stmtGraph.getBranchTargetsOf((JIfStmt) stmt).get(JIfStmt.FALSE_BRANCH_IDX));
+        } else if (stmt instanceof JGotoStmt) {
+          // [ms] bounds are validated in Body if its a valid StmtGraph
+          stmtList.add((Stmt) stmtGraph.getBranchTargetsOf((JGotoStmt) stmt).get(JGotoStmt.BRANCH_IDX));
+        } else if (stmt instanceof JSwitchStmt) {
+          stmtList.addAll(stmtGraph.getBranchTargetsOf((BranchingStmt) stmt));
+        }
+      }
+    }
+
+    for (Trap trap : buildTraps(stmtGraph)) {
+      stmtList.add(trap.getBeginStmt());
+      stmtList.add(trap.getEndStmt());
+      stmtList.add(trap.getHandlerStmt());
+    }
+
+    return stmtList;
   }
 }
