+++ sootup.java.bytecode.frontend/src/test/java/sootup/java/bytecode/frontend/interceptors/TrapTightenerTest.java
@@ -23,6 +23,7 @@ import sootup.core.signatures.MethodSignature;
 import sootup.core.types.ClassType;
 import sootup.core.types.VoidType;
 import sootup.core.util.ImmutableUtils;
+import sootup.core.util.printer.BriefStmtPrinter;
 import sootup.core.util.printer.JimplePrinter;
 import sootup.interceptors.TrapTightener;
 import sootup.java.core.JavaIdentifierFactory;
@@ -35,6 +36,7 @@ import sootup.java.core.views.JavaView;
 @Disabled("FIXME: needs .setTraps() adapted to MutableBlockStmtGraph")
 public class TrapTightenerTest {
   public JimplePrinter jimplePrinter = new JimplePrinter();
+  public final BriefStmtPrinter briefStmtPrinter = new BriefStmtPrinter();
 
   JavaIdentifierFactory factory = JavaIdentifierFactory.getInstance();
   StmtPositionInfo noStmtPositionInfo = StmtPositionInfo.getNoStmtPositionInfo();
@@ -134,7 +136,7 @@ public class TrapTightenerTest {
 
     List<Trap> excepted = new ArrayList<>();
     excepted.add(trap3);
-    List<Trap> actual = jimplePrinter.buildTraps(stmtGraph);
+    List<Trap> actual = briefStmtPrinter.buildTraps(stmtGraph);
     AssertUtils.assertTrapsEquiv(excepted, actual);
   }
   /**
@@ -175,7 +177,7 @@ public class TrapTightenerTest {
 
     List<Trap> excepted = new ArrayList<>();
     excepted.add(trap1);
-    List<Trap> actual = jimplePrinter.buildTraps(stmtGraph);
+    List<Trap> actual = briefStmtPrinter.buildTraps(stmtGraph);
     AssertUtils.assertTrapsEquiv(excepted, actual);
   }
