+++ sootup.java.bytecode.frontend/src/test/java/sootup/java/bytecode/frontend/interceptors/UnreachableCodeEliminatorTest.java
@@ -23,6 +23,7 @@ import sootup.core.types.ClassType;
 import sootup.core.types.PrimitiveType;
 import sootup.core.types.VoidType;
 import sootup.core.util.ImmutableUtils;
+import sootup.core.util.printer.BriefStmtPrinter;
 import sootup.core.util.printer.JimplePrinter;
 import sootup.interceptors.UnreachableCodeEliminator;
 import sootup.java.core.JavaIdentifierFactory;
@@ -35,6 +36,7 @@ import sootup.java.core.views.JavaView;
 public class UnreachableCodeEliminatorTest {
 
   public final JimplePrinter jimplePrinter = new JimplePrinter();
+  public final BriefStmtPrinter briefStmtPrinter = new BriefStmtPrinter();
 
   JavaIdentifierFactory factory = JavaIdentifierFactory.getInstance();
   JavaJimple javaJimple = JavaJimple.getInstance();
@@ -142,7 +144,7 @@ public class UnreachableCodeEliminatorTest {
 
     new UnreachableCodeEliminator().interceptBody(builder, new JavaView(Collections.emptyList()));
 
-    assertEquals(0, jimplePrinter.buildTraps(builder.getStmtGraph()).size());
+    assertEquals(0, briefStmtPrinter.buildTraps(builder.getStmtGraph()).size());
 
     Set<Stmt> expectedStmtsSet = ImmutableUtils.immutableSet(startingStmt, stmt1, ret1);
     AssertUtils.assertSetsEquiv(expectedStmtsSet, builder.getStmtGraph().getNodes());
@@ -178,7 +180,7 @@ public class UnreachableCodeEliminatorTest {
     UnreachableCodeEliminator eliminator = new UnreachableCodeEliminator();
     eliminator.interceptBody(builder, new JavaView(Collections.emptyList()));
 
-    assertEquals(0, jimplePrinter.buildTraps(builder.getStmtGraph()).size());
+    assertEquals(0, briefStmtPrinter.buildTraps(builder.getStmtGraph()).size());
 
     Set<Stmt> expectedStmtsSet = ImmutableUtils.immutableSet(startingStmt, stmt1, ret1);
     assertEquals(expectedStmtsSet, builder.getStmtGraph().getNodes());
